+++ src/components/afishe/regular-day/afishe-regular-day.tsx
